# docs/openapi-production.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PRAXIS OS Product API",
    "description": "Authenticated decision-analysis and simulation beta; not certified operational guidance.",
    "version": "0.9.0"
  },
  "paths": {
    "/v1/decisions/analyze": {
      "post": {
        "summary": "Policy Analysis",
        "operationId": "policy_analysis_v1_decisions_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Context": {
        "properties": {
          "country_code": {
            "type": "string",
            "pattern": "^[A-Za-z]{2,3}$",
            "title": "Country Code"
          },
          "decision_owner_role": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Decision Owner Role",
            "default": "authorized public official"
          },
          "objectives": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "minItems": 1,
            "title": "Objectives"
          },
          "constraints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Constraints"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "country_code",
          "objectives"
        ],
        "title": "Context"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PolicyRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Question"
          },
          "context": {
            "$ref": "#/components/schemas/Context"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/praxis__services__policy_comparison__Option"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 2,
            "title": "Options"
          },
          "weights": {
            "additionalProperties": {
              "type": "number"
            },
            "propertyNames": {
              "enum": [
                "economic",
                "institutional",
                "diplomatic",
                "international_relations",
                "social",
                "fiscal",
                "environmental",
                "security",
                "implementation"
              ]
            },
            "type": "object",
            "maxProperties": 9,
            "minProperties": 1,
            "title": "Weights"
          },
          "scenarios": {
            "items": {
              "$ref": "#/components/schemas/praxis__services__policy_comparison__Scenario"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Scenarios"
          },
          "uncertainty_penalty": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Uncertainty Penalty",
            "default": 0.2
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question",
          "context",
          "options",
          "weights"
        ],
        "title": "PolicyRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "praxis__services__policy_comparison__Criterion": {
        "properties": {
          "dimension": {
            "type": "string",
            "enum": [
              "economic",
              "institutional",
              "diplomatic",
              "international_relations",
              "social",
              "fiscal",
              "environmental",
              "security",
              "implementation"
            ],
            "title": "Dimension"
          },
          "score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": -100.0,
            "title": "Score"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "rationale": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Rationale"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/praxis__services__policy_comparison__Evidence"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Evidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dimension",
          "score",
          "confidence",
          "rationale"
        ],
        "title": "Criterion"
      },
      "praxis__services__policy_comparison__Evidence": {
        "properties": {
          "source": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Source"
          },
          "title": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Title"
          },
          "url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "observed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed At"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "title"
        ],
        "title": "Evidence"
      },
      "praxis__services__policy_comparison__Option": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "title": "Description",
            "default": ""
          },
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/praxis__services__policy_comparison__Criterion"
            },
            "type": "array",
            "maxItems": 9,
            "title": "Criteria"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "Option"
      },
      "praxis__services__policy_comparison__Scenario": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "probability": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Probability"
          },
          "dimension_shocks": {
            "additionalProperties": {
              "type": "number"
            },
            "propertyNames": {
              "enum": [
                "economic",
                "institutional",
                "diplomatic",
                "international_relations",
                "social",
                "fiscal",
                "environmental",
                "security",
                "implementation"
              ]
            },
            "type": "object",
            "title": "Dimension Shocks"
          },
          "option_shocks": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "number"
              },
              "propertyNames": {
                "enum": [
                  "economic",
                  "institutional",
                  "diplomatic",
                  "international_relations",
                  "social",
                  "fiscal",
                  "environmental",
                  "security",
                  "implementation"
                ]
              },
              "type": "object"
            },
            "type": "object",
            "maxProperties": 20,
            "title": "Option Shocks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "probability"
        ],
        "title": "Scenario"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}